more updates mainly for orders

+++ front/placeOrder.php
@@ -18,13 +18,16 @@ $add=$_POST['shipAdd'];
 $shipper=$_POST['shipper'];
 $paymethod=$_POST['payment'];
 $date=date("Y-m-d");
-$sqlo="INSERT INTO `customerorder` VALUES (NULL, '$customerID', '$date', '$amount', '$add', '$shipamt','$paymethod', '$shipper');";
+if($customerID!=0){
+$sqlo="INSERT INTO `customerorders` VALUES (NULL, '$customerID', '$date', '$amount', '$add', '$shipamt','$paymethod', '$shipper');";
 $quer=mysqli_query($con,$sqlo);
 $sql="DELETE FROM `cart` WHERE `CustomerID`='$customerID'";
 $query=mysqli_query($con,$sql);
 
 header("Location:checkout.php?po=1");;
-
+}else{
+	echo "You must be logged in ";
+}
 mysqli_close($con);
 }
 ?>
